move checkout components from the js_modules folder
