+++ src/pages/checkout/event/[event_service_slug].jsx
@@ -11,14 +11,14 @@ import { getDataContentProps } from '../../../utils/file';
 import bc from '../../../common/services/breathecode';
 import useAuth from '../../../common/hooks/useAuth';
 import { isWindow, removeURLParameter, getQueryString, getStorageItem } from '../../../utils';
-import PaymentInfo from '../../../js_modules/checkout/PaymentInfo';
+import PaymentInfo from '../../../common/components/Checkout/PaymentInfo';
 import useSignup from '../../../common/store/actions/signupAction';
 import axiosInstance from '../../../axios';
 import asPrivate from '../../../common/context/PrivateRouteWrapper';
 import LoaderScreen from '../../../common/components/LoaderScreen';
 import useStyle from '../../../common/hooks/useStyle';
-import ServiceSummary from '../../../js_modules/checkout/ServiceSummary';
-import SelectServicePlan from '../../../js_modules/checkout/SelectServicePlan';
+import ServiceSummary from '../../../common/components/Checkout/ServiceSummary';
+import SelectServicePlan from '../../../common/components/Checkout/SelectServicePlan';
 import { ORIGIN_HOST } from '../../../utils/variables';
 import { usePersistentBySession } from '../../../common/hooks/usePersistent';
