+++ src/pages/checkout/index.jsx
@@ -26,20 +26,20 @@ import { getDataContentProps } from '../../utils/file';
 import bc from '../../common/services/breathecode';
 import useAuth from '../../common/hooks/useAuth';
 import useSession from '../../common/hooks/useSession';
-import ContactInformation from '../../js_modules/checkout/ContactInformation';
-import ChooseYourClass from '../../js_modules/checkout/ChooseYourClass';
+import ContactInformation from '../../common/components/Checkout/ContactInformation';
+import ChooseYourClass from '../../common/components/Checkout/ChooseYourClass';
 import { isWindow, getTimeProps, removeURLParameter, getQueryString, getStorageItem, removeStorageItem, slugToTitle, removeSessionStorageItem, getBrowserInfo } from '../../utils';
-import Summary from '../../js_modules/checkout/Summary';
-import PaymentInfo from '../../js_modules/checkout/PaymentInfo';
+import Summary from '../../common/components/Checkout/Summary';
+import PaymentInfo from '../../common/components/Checkout/PaymentInfo';
 import useSignup from '../../common/store/actions/signupAction';
 import axiosInstance from '../../axios';
 import LoaderScreen from '../../common/components/LoaderScreen';
 import ModalInfo from '../../common/components/ModalInfo';
 import useStyle from '../../common/hooks/useStyle';
-import Stepper from '../../js_modules/checkout/Stepper';
-import ServiceSummary from '../../js_modules/checkout/ServiceSummary';
+import Stepper from '../../common/components/Checkout/Stepper';
+import ServiceSummary from '../../common/components/Checkout/ServiceSummary';
 import Text from '../../common/components/Text';
-import SelectServicePlan from '../../js_modules/checkout/SelectServicePlan';
+import SelectServicePlan from '../../common/components/Checkout/SelectServicePlan';
 import { BASE_PLAN, ORIGIN_HOST, BREATHECODE_HOST, currenciesSymbols } from '../../utils/variables';
 import { reportDatalayer } from '../../utils/requests';
 import { getTranslations, processPlans } from '../../common/handlers/subscriptions';
